# docs/swagger/generated.gen.swagger.yml
openapi: 3.0.0
info:
  title: User Management API
  description: ユーザ管理のAPI
  version: 1.0.0
paths:
  /users/{uuid}:
    get:
      operationId: getUser
      description: user取得
      parameters:
        - $ref: '#/components/parameters/UserUUID'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/signup:
    post:
      operationId: UserSignUp
      description: userのsignup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignupResponse'
  /users/signin:
    post:
      operationId: UserSignIn
      description: userのsigninリクエスト
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignInRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignInResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignInError'
components:
  parameters:
    UserUUID:
      name: uuid
      in: path
      description: userのuuid
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  schemas:
    User:
      type: object
      properties:
        uuid:
          type: string
          format: string
          description: userのuuid
        name:
          type: string
          format: string
          description: userの名前
        email:
          type: string
          format: string
          description: userのemail
    UserSignupRequest:
      type: object
      properties:
        name:
          type: string
          format: string
          description: userの名前
        email:
          type: string
          format: string
          description: userのemail
        password:
          type: string
          format: string
          description: userのパスワード
      required:
        - name
        - email
        - password
    UserSignupResponse:
      type: object
      properties:
        uuid:
          type: string
          format: string
          description: userのuuid
      required:
        - uuid
    UserSignInRequest:
      type: object
      properties:
        email:
          type: string
          format: string
          description: userのemail
        password:
          type: string
          format: string
          description: userのパスワード
      required:
        - email
        - password
    UserSignInResponse:
      type: object
      properties:
        uuid:
          type: string
          format: string
          description: userのuuid
      required:
        - uuid
    UserSignInError:
      properties:
        code:
          example: 500-000
          type: string
        errors:
          items:
            example: エラーが発生しました
            type: string
          type: array
      required:
        - code
        - errors
      type: object